fixed PR comments

+++ examples/cim/dqn/components/agent.py
@@ -34,7 +34,7 @@ class CIMAgent(AbsAgent):
             indexes, sample = self._experience_pool.sample_by_key("loss", self._batch_size)
             state = np.asarray(sample["state"])
             action = np.asarray(sample["action"])
-            reward = np.asarray(sample["reward"], dtype=np.float32)
+            reward = np.asarray(sample["reward"])
             next_state = np.asarray(sample["next_state"])
             loss = self._algorithm.train(state, action, reward, next_state)
             self._experience_pool.update(indexes, {"loss": loss})
